requirements.txt
python-dotenv
phidata
groq
googlesearch-python
pycountry

lightweight-charts
yfinance
pandas_ta
ib_insync
nest_asyncio

# finnhub-python